% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_educ_recode_column.R
\name{add_educ_recode_column}
\alias{add_educ_recode_column}
\title{Add column with recoded education to a CenSoc data table
Recode education variable. Based on \url{https://usa.ipums.org/usa-action/variables/EDUC#codes_section}}
\usage{
add_educ_recode_column(df, ...)
}
\arguments{
\item{df}{a CenSoc data frame. Must have `EDUC` column.}
}
\value{
a CenSoc data table with `educ_recode` column.
}
\description{
Add column with recoded education to a CenSoc data table
Recode education variable. Based on \url{https://usa.ipums.org/usa-action/variables/EDUC#codes_section}
}
