% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hispan_recode_column.R
\name{add_hispan_recode_column}
\alias{add_hispan_recode_column}
\title{Add column with recoded hispanic origin to a CenSoc data table}
\usage{
add_hispan_recode_column(df)
}
\arguments{
\item{df}{a CenSoc data frame. Must have `HISPAN` column.}
}
\value{
a CenSoc data frame with `hispan_recode` column.
}
\description{
Recode hispanic variable.
}
