% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_national_censoc.R
\name{create_national_censoc}
\alias{create_national_censoc}
\title{Create a national dataset of matched census and socsec data based on state matched files}
\usage{
create_national_censoc(state_censoc_files)
}
\arguments{
\item{state_censoc_files}{a vector of filenames referring to state matched censoc files}
}
\value{
a dataframe with matched census and socsec data
}
\description{
Create a national dataset of matched census and socsec data based on state matched files
}
