% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_death_proportions_post_samples.R
\name{get_death_proportions_post_samples}
\alias{get_death_proportions_post_samples}
\title{Get posterior samples of death distribution from Bayesian truncated Gompertz model}
\usage{
get_death_proportions_post_samples(mod, cohorts, ncohorts, ngroups, nages)
}
\arguments{
\item{mod}{a JAGS model object}

\item{ncohorts}{number of cohorts}

\item{ngroups}{number of groups}

\item{nages}{number of ages}
}
\value{
an array of posterior samples of death proportions by age, cohort and group
}
\description{
Get posterior samples of death distribution by age, group and cohort and associated uncertainty intervals.
}
