% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deaths_in_jags_format.R
\name{get_deaths_in_jags_format}
\alias{get_deaths_in_jags_format}
\title{Get CenSoc tabulations into data format for JAGS}
\usage{
get_deaths_in_jags_format(tab_df, group_col = NULL)
}
\arguments{
\item{tab_df}{a tabulated CenSoc dataset. Use `tabulate_deaths` on original dataset.}

\item{group_col}{a grouping column. If `NULL` there is no grouping column}
}
\value{
a three-dimensional array with dimensions cohort x age x group level.
}
\description{
Get CenSoc tabulations into data format for JAGS
}
