% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimated_deaths.R
\name{get_estimated_deaths}
\alias{get_estimated_deaths}
\title{Get estimated deaths from Bayesian truncated Gompertz model}
\usage{
get_estimated_deaths(mod, cohorts, ncohorts, group_levels, ngroups, ages, nages,
  alpha_level = 0.05)
}
\arguments{
\item{mod}{a JAGS model object}

\item{cohorts}{vector of cohorts}

\item{ncohorts}{number of cohorts}

\item{group_levels}{vector of groups}

\item{ngroups}{number of groups}

\item{ages}{vector of ages}

\item{nages}{number of ages}

\item{alpha_level}{level of significance. Default is 5\%.}
}
\value{
a dataframe with estimated deaths by age, group and cohort
}
\description{
Get estimated deaths by age, group and cohort and associated uncertainty intervals.
}
