% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimated_hx_ex.R
\name{get_estimated_hx_ex}
\alias{get_estimated_hx_ex}
\title{Get estimates of hazard rates and life expectancy}
\usage{
get_estimated_hx_ex(mod, ncohorts, ngroups, group_labels, nages,
  alpha_level = 0.05)
}
\arguments{
\item{mod}{a JAGS model object}

\item{ncohorts}{number of cohorts}

\item{ngroups}{number of groups}

\item{group_labels}{labels of groups}

\item{nages}{number of ages}

\item{alpha_level}{level of significance. Default is 5\%.}
}
\value{
an array of posterior samples of death proportions by age, cohort and group
}
\description{
Get estimates and uncertainity intervals of mortality indicators
by age, group and cohort and associated uncertainty intervals.
}
