% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_death_distribution.R
\name{get_hmd_death_distribution}
\alias{get_hmd_death_distribution}
\title{HMD death distributions}
\usage{
get_hmd_death_distribution(mx_data, exp_data, ages, remove_na = TRUE,
  single_country)
}
\arguments{
\item{mx_data}{HMD data on death rates}

\item{exp_data}{HMD data on exposure. Must be same dimensions as death rates}

\item{ages}{age range}

\item{remove_na}{whether or not to remove the countries/ years that do not have every age. Default is `TRUE`.}

\item{single_country}{whether or not dataset contains single or multiple countries}
}
\value{
a tibble that contains proportion of male deaths by age
}
\description{
Calculate distribution of deaths for males given death rates and exposure information from HMD
}
