% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hmd_death_pcs.R
\name{get_hmd_death_pcs}
\alias{get_hmd_death_pcs}
\title{Principal components from HMD}
\usage{
get_hmd_death_pcs(dx_data, demeaned = FALSE, single_country, n_pcs = 10)
}
\arguments{
\item{dx_data}{HMD data on death distribution}

\item{demeaned}{whether the values should be demeaned. If \code{TRUE}, the first column of the returned matrix is the mean.}

\item{single_country}{whether or not dataset contains single or multiple countries}

\item{n_pcs}{number of principal components to return. Default is 10.}
}
\value{
a matrix that contains principal components by age of logit death distributions.
}
\description{
Calculate principal components (on logit scale) based on death distributions from HMD.
}
