% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_data.R
\name{get_jags_data}
\alias{get_jags_data}
\title{Get left and right truncation ages for each cohort and grouping variable level}
\usage{
get_jags_data(tab_df, ncohorts, ngroups, min_age)
}
\arguments{
\item{tab_df}{A tabulated CenSoc dataframe}

\item{ncohorts}{Number of cohorts}

\item{ngroups}{Number of group levels}

\item{min_age}{The minimum age of estimation, i.e. where to start estimates of dx from (lowest that makes sense is probably 40)}
}
\value{
a list containing data elements in the right form for JAGS input
}
\description{
Given a tabulated deaths data frame and the number of groups and cohorts,
will return list containing data elements in the right form for JAGS input
}
