% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_data.R
\name{get_meta_data}
\alias{get_meta_data}
\title{Get meta data associated with tabulated deaths by cohort, age of death and grouping variable}
\usage{
get_meta_data(tab_df, min_age)
}
\arguments{
\item{tab_df}{A tabulated CenSoc dataframe}

\item{min_age}{The minimum age of estimation, i.e. where to start estimates of dx from (lowest that makes sense is probably 40)}
}
\value{
a list of metadata
}
\description{
Given a tabulated deaths data frame and a minimum age of estimation,
will return age vector and length, cohort length, group levels and length
}
