% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truncation_ages.R
\name{get_truncation_ages}
\alias{get_truncation_ages}
\title{Get left and right truncation ages for each cohort and grouping variable level}
\usage{
get_truncation_ages(tab_df, ncohorts, ngroups)
}
\arguments{
\item{tab_df}{A tabulated CenSoc dataframe}

\item{ncohorts}{Number of cohorts}

\item{ngroups}{Number of group levels}
}
\value{
a list containing matrices of left and right truncation ages
}
\description{
Given a tabulated deaths data frame and the number of groups and cohorts,
will return matrices containing left and right truncation ages
}
