% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_census.R
\name{load_census}
\alias{load_census}
\title{Load a census state file}
\usage{
load_census(state_file, cols_to_keep)
}
\arguments{
\item{state_file}{string indicating location of file}

\item{cols_to_keep}{vector indicating colname names of variables to keep}
}
\description{
Load a census state file
}
