% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ex.R
\name{plot_ex}
\alias{plot_ex}
\title{Plot life expectancy by cohort and group}
\usage{
plot_ex(ex_df, ex_age, group_title)
}
\arguments{
\item{ex_df}{dataframe containing life expectancy estimates and uncertainty for each cohort, age and group}

\item{ex_age}{age for life expectancy to be plotted. Minimum is the minimum age of estimation (default is 55).}

\item{group_title}{string indicating the type of group being considered, e.g. "Education"}
}
\value{
a ggplot object
}
\description{
Plots life expectancy estimates and uncertainty over cohort for each group.
}
