% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hmd.R
\name{read_hmd}
\alias{read_hmd}
\title{Read in HMD text file}
\usage{
read_hmd(file_name, add_country = FALSE, country_name = NULL)
}
\arguments{
\item{file_name}{string indicating location of hmd text file}

\item{add_country}{logicial indicating whether or not to add a column with the country name. Default is `FALSE`}

\item{country_name}{country name}
}
\value{
a data table with the contents of HMD file read in.
}
\description{
Read in HMD text file
}
