% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gompertz_model.R
\name{run_gompertz_model}
\alias{run_gompertz_model}
\title{Run MCMC for Bayesian trunacted Gompertz model}
\usage{
run_gompertz_model(tab_df, min_age, number_chains = 4,
  number_iterations = 5000)
}
\arguments{
\item{tab_df}{A tabulated CenSoc dataframe}

\item{min_age}{The minimum age of estimation, i.e. where to start estimates of dx from (lowest that makes sense is probably 40)}

\item{ncohorts}{Number of cohorts}

\item{ngroups}{Number of group levels}
}
\value{
a list containing data elements in the right form for JAGS input
}
\description{
Given a tabulated deaths data frame and the number of groups and cohorts,
will return list containing data elements in the right form for JAGS input
}
