% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_deaths.R
\name{tabulate_deaths}
\alias{tabulate_deaths}
\title{Calculate number of deaths by birth year, age of death and possible grouping variables}
\usage{
tabulate_deaths(df, min_age = 60, ...)
}
\arguments{
\item{df}{CenSoc dataframe}

\item{min_age}{the minimum age of death in CenSoc to consider. Default is 60. Shouldn't be any less than 55 as coverage becomes low.}

\item{...}{grouping variables (in addition to year and race)}
}
\value{
tabulated dataframe
}
\description{
Given a CenSoc data frame with individual birth years, ages of death and other grouping variables,
will return tabulated deaths by group
}
