% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_gompertz_model.R
\name{truncated_gompertz_model}
\alias{truncated_gompertz_model}
\title{Run MCMC for Bayesian truncated Gompertz model}
\usage{
truncated_gompertz_model(tab_df, min_age, number_chains = 3,
  number_iterations = 7500, run_update = FALSE)
}
\arguments{
\item{tab_df}{A tabulated CenSoc dataframe}

\item{min_age}{The minimum age of estimation, i.e. where to start estimates of dx from (lowest that makes sense is probably 40)}

\item{number_chains}{The number of chains of MCMC}

\item{number_iterations}{The number of iterations per chain of MCMC}

\item{run_update}{Whether or not to automatically run model update if model did not converge}
}
\value{
a JAGS model object
}
\description{
Estimates a Bayesian truncated Gompertz model based on an inputted tabulated dataframe of deaths
and a minimum age of estimation.
}
