% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_deaths.R
\name{tabulate_deaths}
\alias{tabulate_deaths}
\title{Calculate number of deaths by birth year, age of death and possible grouping variables}
\usage{
tabulate_deaths(df, ...)
}
\arguments{
\item{df}{CenSoc datafrane}

\item{...}{grouping variables (in addition to year and race)}
}
\value{
tabulated dataframe
}
\description{
Given a CenSoc data frame with individual birth years, ages of death and other grouping variables,
will return tabulated deaths by group
}
